requirements/base.txt
fastapi==0.103.2
fastapi-utils==0.7.0
aiofiles==0.7.0
chroma-hnswlib==0.7.6
chromadb==0.5.5
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
langchain==0.2.10
langchain-chroma==0.1.2
langchain-community==0.2.10
langchain-core==0.2.32
langchain-openai==0.1.21
langchain-text-splitters==0.2.2
langsmith==0.1.99
Markdown==3.6
matplotlib==3.8.3
openai==1.40.6
pandas==2.1.3
pdfplumber==0.11.3
pipenv==2023.7.23
pydantic==2.7.4
pydot==3.0.1
python-dotenv==1.0.0
scikit-learn==1.3.2
SQLAlchemy==2.0.23
sqlparse==0.4.4
tiktoken==0.7.0
tinycss2==1.2.1
tokenizers==0.20.0
torch==2.2.0
tqdm==4.66.5
uvicorn==0.24.0.post1
python-multipart==0.0.9
elasticsearch==7.17.0
pytest==8.3.2
pytest-asyncio==0.23.8